[수정] person api요청, dom 조작 두 기능으로 파일을 나누고 모듈화
