+++ kimyongyeon/HelloSpringStudy/src/main/resources/static/public/index.html
@@ -23,6 +23,8 @@
                 </tr>
             </table>
         </div>
-        <script src="script.js"></script>
+        <script src="main.js" type="module"></script>
+        <script src="personApi.js" type="module"></script>
+        <script src="domUtil.js" type="module"></script>
     </body>
 </html>
