+++ kimyongyeon/HelloSpringStudy/src/main/resources/static/public/main.js
@@ -1,0 +1,9 @@
+import {fetchPersonData} from "./personApi.js";
+import {createPersonDataTable} from "./domUtil.js";
+
+const displayPersonData = async () => {
+        const data = await fetchPersonData();
+        createPersonDataTable(data);
+};
+
+displayPersonData()
